Better estimation of ball position
